Warnings: Reduce the amount of warnings

+++ src/tools/ToolsTBC/vmap_tools/vmap4_extractor/vmapexport.cpp
@@ -87,7 +87,7 @@ void ReadLiquidTypeTableDBC()
 {
     printf("Read LiquidType.dbc file...");
     DBCFile dbc("DBFilesClient\\LiquidType.dbc");
-    if(!dbc.open())
+    if (!dbc.open())
     {
         printf("Fatal error: Invalid LiquidType.dbc file format!\n");
         exit(1);
@@ -161,13 +161,13 @@ bool ExtractSingleWmo(std::string& fname)
     bool file_ok = true;
     std::cout << "Extracting " << fname << std::endl;
     WMORoot froot(fname);
-    if(!froot.open())
+    if (!froot.open())
     {
         printf("Couldn't open RootWmo!!!\n");
         return true;
     }
     FILE *output = fopen(szLocalFile,"wb");
-    if(!output)
+    if (!output)
     {
         printf("couldn't open %s for writing!\n", szLocalFile);
         return false;
@@ -188,7 +188,7 @@ bool ExtractSingleWmo(std::string& fname)
 
             std::string s = groupFileName;
             WMOGroup fgroup(s);
-            if(!fgroup.open())
+            if (!fgroup.open())
             {
                 printf("Could not open all Group file for: %s\n", plain_name);
                 file_ok = false;
@@ -220,7 +220,7 @@ void ParsMapFiles()
         sprintf(id, "%04u", map_ids[i].id);
         sprintf(fn,"World\\Maps\\%s\\%s.wdt", map_ids[i].name, map_ids[i].name);
         WDTFile WDT(fn, map_ids[i].name);
-        if(WDT.init(id, map_ids[i].id))
+        if (WDT.init(id, map_ids[i].id))
         {
             printf("Processing Map %u\n[", map_ids[i].id);
             for (int x = 0; x < 64; ++x)
@@ -268,9 +268,9 @@ bool scan_patches(char* scanmatch, std::vector<std::string>& pArchiveNames)
             sprintf(path, "%s.MPQ", scanmatch);
         }
 #ifdef __linux__
-        if(FILE* h = fopen64(path, "rb"))
+        if (FILE* h = fopen64(path, "rb"))
 #else
-        if(FILE* h = fopen(path, "rb"))
+        if (FILE* h = fopen(path, "rb"))
 #endif
         {
             fclose(h);
@@ -284,7 +284,7 @@ bool scan_patches(char* scanmatch, std::vector<std::string>& pArchiveNames)
 
 bool fillArchiveNameVector(std::vector<std::string>& pArchiveNames)
 {
-    if(!hasInputPathParam)
+    if (!hasInputPathParam)
         getGamePath();
 
     printf("\nGame path: %s\n", input_path);
@@ -345,13 +345,13 @@ bool fillArchiveNameVector(std::vector<std::string>& pArchiveNames)
     {
         printf("Locale: %s\n", i->c_str());
         sprintf(path, "%s%s/patch-%s", input_path, i->c_str(), i->c_str());
-        if(scan_patches(path, pArchiveNames))
+        if (scan_patches(path, pArchiveNames))
             foundOne = true;
     }
 
     printf("\n");
 
-    if(!foundOne)
+    if (!foundOne)
     {
         printf("no locale found\n");
         return false;
@@ -368,13 +368,13 @@ bool processArgv(int argc, char ** argv, const char *versionString)
 
     for(int i = 1; i < argc; ++i)
     {
-        if(strcmp("-s",argv[i]) == 0)
+        if (strcmp("-s",argv[i]) == 0)
         {
             preciseVectorData = false;
         }
-        else if(strcmp("-d",argv[i]) == 0)
+        else if (strcmp("-d",argv[i]) == 0)
         {
-            if((i+1)<argc)
+            if ((i+1)<argc)
             {
                 hasInputPathParam = true;
                 strncpy(input_path, argv[i + 1], sizeof(input_path));
@@ -389,11 +389,11 @@ bool processArgv(int argc, char ** argv, const char *versionString)
                 result = false;
             }
         }
-        else if(strcmp("-?",argv[1]) == 0)
+        else if (strcmp("-?",argv[1]) == 0)
         {
             result = false;
         }
-        else if(strcmp("-l",argv[i]) == 0)
+        else if (strcmp("-l",argv[i]) == 0)
         {
             preciseVectorData = true;
         }
@@ -403,7 +403,7 @@ bool processArgv(int argc, char ** argv, const char *versionString)
             break;
         }
     }
-    if(!result)
+    if (!result)
     {
         printf("Extract %s.\n",versionString);
         printf("%s [-?][-s][-l][-d <path>]\n", argv[0]);
@@ -491,9 +491,9 @@ int main(int argc, char ** argv)
             printf("FATAL ERROR: Map.dbc not found in data file.\n");
             return 1;
         }
-        map_count = dbc->getRecordCount();
-        map_ids=new map_id[map_count];
-        for (unsigned int x=0;x<map_count;++x)
+        map_count = static_cast<uint32_t>(dbc->getRecordCount());
+        map_ids = new map_id[map_count];
+        for (unsigned int x = 0; x < map_count; ++x)
         {
             map_ids[x].id = dbc->getRecord(x).getUInt(0);
 
@@ -516,7 +516,7 @@ int main(int argc, char ** argv)
         ParsMapFiles();
         delete [] map_ids;
         //nError = ERROR_SUCCESS;
-        // Extract models, listed in DameObjectDisplayInfo.dbc
+        // Extract models, listed in GameObjectDisplayInfo.dbc
         ExtractGameobjectModels();
     }
